myproject/requirement.txt
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Pillow-10.4.0